+++ source/Public/New-DscResourcePowerShellHelp.ps1
@@ -1,3 +1,4 @@
+
 <#
     .SYNOPSIS
         New-DscResourcePowerShellHelp generates PowerShell compatible help files