+++ source/Public/New-DscResourceWikiPage.ps1
@@ -1,3 +1,4 @@
+
 <#
     .SYNOPSIS
         New-DscResourceWikiPage generates wiki pages that can be uploaded to GitHub