+++ CHANGELOG.md
@@ -24,6 +24,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - DscResource.DocGenerator
   - Updated pipeline files to support resolving dependencies using ModuleFast
     or PSResourceGet.
+  - Add blank line at top of each public command source file for PlatyPS
+    to work. ModuleBuilder has no option to add a prefix per source file,
+    the only alternative would be to move the comment-based help inside the
+    function-block which is currently against the style guideline.
 
 ### Fixed